feat: needs first editor, needs reference checker etc
